Split CodexView's reactive layer into per-tab view classes (#1737)
CodexView had grown into one 820-line class holding the Enemies/Zones/Skills
tabs' selection state, ~30 deriveds and the deep-link constructor together.
Split the per-tab reactive logic into EnemiesTabView, ZonesTabView and
SkillsTabView (enemies-tab-view.svelte.ts, zones-tab-view.svelte.ts,
skills-tab-view.svelte.ts), mirroring the screen's existing per-tab component
split. CodexView is now a thin orchestrator that owns only the cross-tab
concerns (active tab, the once-fetched stats/challenge-error state, the
shared statistics query engine, and cross-dossier links) and delegates every
other property/method to the owning tab's instance, so its public API - and
every consuming component and the existing test suite - is unchanged.

Also moved the shared catalogue filters (liveEnemies/liveZones/liveSkills)
into codex-display.ts so all three tab views can reuse them without an
inter-tab import.

No behavior change - the existing codex-view.svelte.test.ts suite (unmodified)
pins the seams.
